fix: resolve security vulnerabilities found by gosec
- Fix integer overflow vulnerabilities (CWE-190)
  - Convert bytes to uint16 directly in request.go
  - Add safe conversion from int to uint64 in forwarder.go
  - Fix uint16 port conversions in net.go and socks.go

- Mitigate path traversal vulnerabilities (CWE-22)
  - Add path sanitization using filepath.Clean() for config file access
  - Improve certificate file path handling

- Prevent Slowloris attack vulnerability (CWE-400)
  - Configure ReadHeaderTimeout in HTTP server

- Restrict file permissions (CWE-276)
  - Set log file permissions to 0600 instead of 0644

+++ internal/socks/request.go
@@ -254,7 +254,7 @@ func readAddrSpec(r io.Reader) (*AddrSpec, error) {
 	if _, err := io.ReadAtLeast(r, port, 2); err != nil {
 		return nil, err
 	}
-	d.Port = uint16((int(port[0]) << 8) | int(port[1]))
+	d.Port = (uint16(port[0]) << 8) | uint16(port[1])
 
 	return d, nil
 }
